add more benchmarks

+++ Program.cs
@@ -1,14 +1,17 @@
-﻿
-using Benchmark_UUID_ULID_SqlServer;
-using Benchmark_UUID_ULID_SqlServer.Data;
+﻿using Benchmark_UUID_ULID_SqlServer.Data;
 using BenchmarkDotNet.Running;
 using Microsoft.EntityFrameworkCore;
 
-using (BenchmarkContext context = new BenchmarkContext())
-{
-    context.Database.EnsureDeleted();
-    context.Database.Migrate();
-}
-//BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
-var summary = BenchmarkRunner
-    .Run<UuidVsUlidSqlInsert>();
+Console.WriteLine("Soll die DB neu erstellt werden J/N:");
+if (Console.ReadKey().KeyChar == 'J')
+    // Recreate DB
+    using (BenchmarkContext context = new BenchmarkContext())
+    {
+        context.Database.EnsureDeleted();
+        context.Database.Migrate();
+    }
+
+// Run Benchmarks
+BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
+//var summary = BenchmarkRunner
+//    .Run<UuidVsUlidSqlInsert>();
